move version detection out of codegen
